+++ beego/restful/utm_test.go
@@ -1,4 +1,4 @@
-package utm
+package restful
 
 import (
 	"net/url"
